# studymanager/src/main/resources/openapi/StudyManagerAPI.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: StudyManager API
security:
  - OAuth:
      - read
      - write
servers:
  - url: https://localhost:8080/api/v1
    description: Local Test Server
paths:
...
  /components/{componentType}/{componentId}/api/{slug}:
    parameters:
      - $ref: '#/components/parameters/ComponentType'
      - name: componentId
        in: path
        schema:
          type: string
        required: true
      - name: slug
        in: path
        schema:
          type: string
        required: true
    post:
      tags:
        - components
      operationId: accessModuleSpecificEndpoint
      description: access module specific endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Endpoint not found
        '500':
          description: Error
...
  /studies/{studyId}/observations/{observationId}/tokens:
    parameters:
      - $ref: '#/components/parameters/StudyId'
      - $ref: '#/components/parameters/ObservationId'

    post:
      tags:
        - observations
      description: Create and add observation endpoint token
      operationId: createToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndpointToken'
      responses:
        '201':
          description: Token successfully added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointToken'
        '400':
          description: bad request
        '404':
          description: not found

    get:
      tags:
        - observations
      description: Get all observation endpoint tokens
      operationId: getTokens
      responses:
        '200':
          description: Tokens successfully returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EndpointToken'
        '404':
          description: not found
  /studies/{studyId}/observations/{observationId}/tokens/{tokenId}:
    parameters:
      - $ref: '#/components/parameters/StudyId'
      - $ref: '#/components/parameters/ObservationId'
      - $ref: '#/components/parameters/TokenId'

    get:
      tags:
        - observations
      description: Get observation endpoint token
      operationId: getToken
      responses:
        '201':
          description: Token successfully returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointToken'
        '404':
          description: not found

    put:
      tags:
        - observations
      description: Update label from observation endpoint token
      operationId: updateTokenLabel
      parameters:
        - $ref: '#/components/parameters/StudyId'
        - $ref: '#/components/parameters/ObservationId'
        - $ref: '#/components/parameters/TokenId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndpointToken'
      responses:
        '200':
          description: Token label successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointToken'

        '400':
          description: could not update observation
        '404':
          description: not found

    delete:
      tags:
        - observations
      description: Delete observation endpoint token
      operationId: deleteToken
      responses:
        '204':
          description: Token successfully deleted
        '404':
          description: not found
...
  /studies/{studyId}/interventions/{interventionId}/trigger:
    parameters:
      - $ref: '#/components/parameters/StudyId'
      - $ref: '#/components/parameters/InterventionId'
    get:
      tags:
        - interventions
      description: Get trigger
      operationId: getTrigger
      responses:
        '200':
          description: trigger returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trigger'
        '400':
          description: bad request
    put:
      tags:
        - interventions
      description: Update trigger
      operationId: updateTrigger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Trigger'
      responses:
        '200':
          description: trigger updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trigger'
        '400':
          description: bad request
    post:
      tags:
        - interventions
      description: Trigger intervention for specific participants
      operationId: triggerIntervention
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerRequest'
      responses:
        '202':
          description: Trigger accepted
        '400':
          description: Bad request
...
  /studies/import/study:
    post:
      tags:
        - importExport
      description: Import Study
      operationId: importStudy
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary

      responses:
        '201':
          description: import was successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Study'
        '400':
          description: bad request
  /studies/{studyId}/export/study:
    parameters:
      - $ref: '#/components/parameters/StudyId'
    get:
      tags:
        - importExport
      description: Export Study
      operationId: exportStudy
      responses:
        '200':
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyImportExport'
        '400':
          description: bad request
...
  /auditlog/study/{studyId}:
    parameters:
      - $ref: '#/components/parameters/StudyId'
    get:
      tags:
        - auditLog
      description: Provides metadata of the audit logs for the requested study
      operationId: getAuditLogMetadata
      responses:
        '200':
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogMetadata'
        '404':
          description: not found
...
components:
  schemas:
...
    Study:
      type: object
      properties:
        studyId:
          $ref: '#/components/schemas/StudyId'
        title:
          type: string
        purpose:
          type: string
        participantInfo:
          type: string
        consentInfo:
          type: string
        duration:
          $ref: '#/components/schemas/StudyDuration'
        finishText:
          type: string
        status:
          $ref: '#/components/schemas/StudyStatus'
        start:
          type: string
          format: date
          readOnly: true
        end:
          type: string
          format: date
          readOnly: true
        plannedStart:
          type: string
          format: date
        plannedEnd:
          type: string
          format: date
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
        applicationAccess:
          type: array
          description: Application types which the user have access to
          items:
            type: string
        userRoles:
          type: array
          description: the roles the _current user_ has within _this_ study.
          items:
            $ref: '#/components/schemas/StudyRole'
          uniqueItems: true
          readOnly: true
          minItems: 1
        contact:
          $ref: '#/components/schemas/Contact'
    Contact:
      type: object
      properties:
        institute:
          type: string
        person:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
...
    StudyStatus:
      type: string
      enum:
        - draft
        - preview
        - paused-preview
        - active
        - paused
        - closed
      default: draft
    StudyGroup:
      type: object
      properties:
        studyId:
          $ref: '#/components/schemas/StudyId'
        studyGroupId:
          $ref: '#/components/schemas/Id'
        title:
          type: string
        purpose:
          type: string
        duration:
          $ref: '#/components/schemas/StudyDuration'
        numberOfParticipants:
          type: integer
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
    ObservationGroup:
      type: object
      properties:
        studyId:
          $ref: '#/components/schemas/StudyId'
        observationGroupId:
          $ref: '#/components/schemas/Id'
        title:
          type: string
        purpose:
          type: string
        numberOfParticipants:
          type: integer
          readOnly: true
        numberOfObservations:
          type: integer
          readOnly: true
        numberOfInterventions:
          type: integer
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
      required:
        - title
    StudyDuration:
      type: object
      properties:
        value:
          type: integer
        unit:
          type: string
          description: unit of time
          enum:
            - MINUTE
            - HOUR
            - DAY
...
    Observation:
      type: object
      properties:
        studyId:
          $ref: '#/components/schemas/StudyId'
        observationId:
          $ref: '#/components/schemas/Id'
        studyGroupId:
          $ref: '#/components/schemas/IdReference'
        title:
          type: string
        purpose:
          type: string
        participantInfo:
          type: string
        type:
          type: string
        properties:
          type: object
          additionalProperties: true
        schedule:
          oneOf:
            - $ref: './Event.yaml/#/components/schemas/Event'
            - $ref: './RelativeEvent.yaml/#/components/schemas/RelativeEvent'
          discriminator:
            propertyName: type
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
        hidden:
          type: boolean
        noSchedule:
          type: boolean
          default: false
        reminder:
          type: boolean
          default: false
        observationGroupIds:
          type: array
          description: The set of observation group Ids the observation is part of
          items:
            $ref: '#/components/schemas/IdReference'
          uniqueItems: true
...
    TriggerRequest:
      type: object
      required:
        - participantIds
      properties:
        participantIds:
          type: array
          items:
            type: integer
          description: List of participant IDs to trigger
    EndpointToken:
      type: object
      properties:
        tokenId:
          $ref: '#/components/schemas/Id'
        tokenLabel:
          type: string
        created:
          type: string
          format: date-time
          readOnly: true
        token:
          type: string
          readOnly: true
      required:
        - tokenId
        - tokenLabel
        - created
        - token
...
    Intervention:
      type: object
      properties:
        studyId:
          $ref: '#/components/schemas/StudyId'
        interventionId:
          $ref: '#/components/schemas/Id'
        studyGroupId:
          $ref: '#/components/schemas/IdReference'
        title:
          type: string
        purpose:
          type: string
        schedule:
          oneOf:
            - $ref: './Event.yaml/#/components/schemas/Event'
            - $ref: './RelativeEvent.yaml/#/components/schemas/RelativeEvent'
          discriminator:
            propertyName: type
        trigger:
          $ref: '#/components/schemas/Trigger'
        actions:
          type: array
          items:
            $ref: '#/components/schemas/Action'
        observationGroupIds:
          type: array
          description: The set of observation group Ids the intervention is part of
          items:
            $ref: '#/components/schemas/IdReference'
          uniqueItems: true
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
    Trigger:
      type: object
      readOnly: true
      properties:
        type:
          type: string
        properties:
          type: object
          additionalProperties: true
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
    Action:
      type: object
      properties:
        actionId:
          $ref: '#/components/schemas/Id'
        type:
          type: string
        properties:
          type: object
          additionalProperties: true
        created:
          type: string
          format: date-time
          readOnly: true
        modified:
          type: string
          format: date-time
          readOnly: true
    StudyId:
      type: integer
      format: int64
      readOnly: true
    Id:
      type: integer
      format: int32
      readOnly: true
    IdReference:
      type: integer
      format: int32
    StudyImportExport:
      type: object
      properties:
        study:
          $ref: '#/components/schemas/Study'
        studyGroups:
          type: array
          items:
            $ref: '#/components/schemas/StudyGroup'
        observationGroups:
          type: array
          items:
            $ref: '#/components/schemas/ObservationGroup'
        observations:
          type: array
          items:
            $ref: '#/components/schemas/Observation'
        interventions:
          type: array
          items:
            $ref: '#/components/schemas/Intervention'
        participants:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantInfo'
        integrations:
          type: array
          items:
            $ref: '#/components/schemas/IntegrationInfo'
    ParticipantInfo:
      type: object
      properties:
        studyGroup:
          type: integer
        observationGroups:
          type: array
          items:
            type: integer
          uniqueItems: true
    IntegrationInfo:
      type: object
      properties:
        name:
          type: string
        observationId:
          $ref: '#/components/schemas/Id'
...
    StudyRole:
      description: |
        We differentiate between the following roles within a study:
        * `STUDY_ADMIN` (_Study Administrator_): **All** rights to create and manage studies, configure components and assign participants, invite and manage operators and viewers and see study data.
        * `STUDY_OPERATOR` (_Study Operator_): Rights to manage studies, configure components and assign participants, **no rights to see data**.
        * `STUDY_VIEWER` (_Study Viewer_): Rights to see study daƒta only.
      type: string
      enum:
        - STUDY_ADMIN
        - STUDY_OPERATOR
        - STUDY_VIEWER
      externalDocs:
        url: https://more-platform.atlassian.net/browse/MORE-213
...
    AuditLogMetadata:
      description: |
        Gives back current length of audit log entries with an integer and export format
      type: object
      properties:
        studyId:
          $ref: '#/components/schemas/StudyId'
        length:
          type: integer
          format: int64
...
  parameters:
    ComponentType:
      name: componentType
      in: path
      schema:
        type: string
        enum:
          - action
          - trigger
          - observation
      required: true
    StudyId:
      name: studyId
      in: path
      schema:
        $ref: '#/components/schemas/StudyId'
      required: true
...
    ObservationId:
      name: observationId
      in: path
      schema:
        type: integer
        format: int32
      required: true
    InterventionId:
      name: interventionId
      in: path
      schema:
        type: integer
        format: int32
      required: true
...
    TokenId:
      name: tokenId
      in: path
      schema:
        type: integer
        format: int32
      required: true
...
  securitySchemes:
    OAuth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://auth.more.redlink.io/authorize #TODO
          scopes:
            read: read studies
            write: write studies